% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.r
\name{circumference_for_age}
\alias{circumference_for_age}
\title{Head circumference for age}
\usage{
circumference_for_age(circumference, age, gender = "male",
  unit.length = "in")
}
\description{
Head circumference for age
}

