% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/growthcharts-package.r
\docType{package}
\name{growthcharts}
\alias{growthcharts}
\alias{growthcharts-package}
\title{WHO growth charts}
\description{
Generate growth charts based on WHO data.
}
\author{
Aaron Wolen \email{aaron@wolen.com}

Maintainer: Aaron Wolen \email{aaron@wolen.com}
}

