% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/growthcharts-package.r
\docType{data}
\name{growthdata}
\alias{growthdata}
\title{WHO Growth Standards}
\description{
Growth charts are generated using summary statistics calculated using the
World Health Organization's Child Growth Standards and made available by the
Centers for Disease Control.
}
\references{
WHO Multicentre Growth Reference Study Group. \href{http://www.who.int/entity/childgrowth/standards/Technical_report.pdf}{WHO Child Growth Standards: Length/height-for-age, weight-for-age, weight-for-length, weight-for-height and body mass index-for-age: Methods and development}. Geneva: World Health Organization, 2006
}

