% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.r
\name{length_for_age}
\alias{length_for_age}
\title{Length for age}
\usage{
length_for_age(length = NULL, age = NULL, gender = "male",
  units = c("cm", "month"))
}
\arguments{
\item{length}{child's length}

\item{age}{age in months at the time of trait measurment}

\item{gender}{(\code{male} or \code{female})}

\item{units}{character vector indicating the units of measures used for the
plotted traits}
}
\description{
Length for age
}

