% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.r
\name{weight_for_age}
\alias{weight_for_age}
\title{Weight for age}
\usage{
weight_for_age(weight = NULL, age = NULL, gender = "male",
  units = c("kg", "month"))
}
\arguments{
\item{age}{age in months at the time of trait measurment}

\item{gender}{(\code{male} or \code{female})}

\item{units}{character vector indicating the units of measures used for the
plotted traits}
}
\description{
Weight for age
}

