% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.r
\name{weight_for_length}
\alias{weight_for_length}
\title{Weight for length}
\usage{
weight_for_length(weight = NULL, length = NULL, gender = "male",
  units = c("kg", "cm"), palette = "Greys")
}
\arguments{
\item{weight}{child's weight}

\item{length}{child's length}

\item{gender}{(\code{male} or \code{female})}

\item{units}{character vector indicating the units of measures used for the
plotted traits}

\item{palette}{color brewer palette applied to percentiles}
}
\description{
Weight for length
}

