% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.r
\name{circumference_for_age}
\alias{circumference_for_age}
\title{Head circumference for age}
\usage{
circumference_for_age(circumference = NULL, age = NULL, gender = "male",
  units = c("cm", "month"), color = "blue", palette = "Greys")
}
\arguments{
\item{circumference}{head circumference}

\item{age}{age in months at the time of trait measurment}

\item{gender}{(\code{male} or \code{female})}

\item{units}{character vector indicating the units of measures used for the
plotted traits}

\item{color}{color applied to plotted points}

\item{palette}{color brewer palette applied to percentiles}
}
\description{
Head circumference for age
}

