% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/growthcharts-package.r
\docType{package}
\name{growthcharts}
\alias{growthcharts}
\alias{growthcharts-package}
\title{Growth charts}
\description{
Growth charts for infants and children ages 0 to 2 using the World Health
Organization's child growth standards.
}
\author{
Aaron Wolen \email{aaron@wolen.com}

Maintainer: Aaron Wolen \email{aaron@wolen.com}
}

